chore: Undo v2 split and overwrite core engine natively
